Created Enemy, Adjusted environment and materials, tested for win
Created an enemy which moves by waypoint system. They cause damage to the player when they collide.  If the player bumps the wall they will take damage.
Adjusted the colors for the level geometry including changing to Unlit shader.
Made a WebGL build and play tested a bit.  I was able to beat the game.