Created a LevelManager
Created a  LevelManager which is in charge of loading the current or next level scene.  No more death message, the player will spawn into the last Section of the level instead.

Created a silly damage text but it doesn't give a value of damage but rather random words like Ouch, WTF!, etc